@@ -1,54 +1,58 @@
 {
-    "name": "livewire/livewire",
-    "description": "A front-end framework for Laravel.",
-    "license": "MIT",
-    "authors": [
-        {
-            "name": "Caleb Porzio",
-            "email": "calebporzio@gmail.com"
-        }
+  "name": "livewire/livewire",
+  "description": "A front-end framework for Laravel.",
+  "license": "MIT",
+  "authors": [
+    {
+      "name": "Caleb Porzio",
+      "email": "calebporzio@gmail.com"
+    }
+  ],
+  "require": {
+    "php": "^8.1",
+    "illuminate/database": "^10.0|^11.0",
+    "illuminate/routing": "^10.0|^11.0",
+    "illuminate/support": "^10.0|^11.0",
+    "illuminate/validation": "^10.0|^11.0",
+    "league/mime-type-detection": "^1.9",
+    "symfony/console": "^6.0|^7.0",
+    "symfony/http-kernel": "^6.2|^7.0",
+    "laravel/prompts": "^0.1.24|^0.2|^0.3"
+  },
+  "require-dev": {
+    "psy/psysh": "^0.11.22|^0.12",
+    "mockery/mockery": "^1.3.1",
+    "phpunit/phpunit": "^10.4",
+    "laravel/framework": "^10.15.0|^11.0",
+    "orchestra/testbench": "^8.21.0|^9.0",
+    "orchestra/testbench-dusk": "^8.24|^9.1",
+    "calebporzio/sushi": "^2.1"
+  },
+  "autoload": {
+    "files": [
+      "src/helpers.php"
     ],
-    "require": {
-        "php": "^7.2.5|^8.0",
-        "symfony/http-kernel": "^5.0|^6.0",
-        "illuminate/support": "^7.0|^8.0|^9.0|^10.0",
-        "illuminate/database": "^7.0|^8.0|^9.0|^10.0",
-        "illuminate/validation": "^7.0|^8.0|^9.0|^10.0",
-        "league/mime-type-detection": "^1.9"
-    },
-    "require-dev": {
-        "psy/psysh": "@stable",
-        "mockery/mockery": "^1.3.1",
-        "phpunit/phpunit": "^8.4|^9.0",
-        "laravel/framework": "^7.0|^8.0|^9.0|^10.0",
-        "orchestra/testbench": "^5.0|^6.0|^7.0|^8.0",
-        "orchestra/testbench-dusk": "^5.2|^6.0|^7.0|^8.0",
-        "calebporzio/sushi": "^2.1"
-    },
-    "autoload": {
-        "files": [
-            "src/helpers.php"
-        ],
-        "psr-4": {
-            "Livewire\\": "src/"
-        }
-    },
-    "autoload-dev": {
-        "psr-4": {
-            "App\\": "vendor/orchestra/testbench-core/laravel/app",
-            "Tests\\": "tests/"
-        }
-    },
-    "extra": {
-        "laravel": {
-            "providers": [
-                "Livewire\\LivewireServiceProvider"
-            ],
-            "aliases": {
-                "Livewire": "Livewire\\Livewire"
-            }
-        }
-    },
-    "minimum-stability": "dev",
-    "prefer-stable": true
+    "psr-4": {
+      "Livewire\\": "src/"
+    }
+  },
+  "autoload-dev": {
+    "psr-4": {
+      "App\\": "vendor/orchestra/testbench-core/laravel/app",
+      "Tests\\": "tests/",
+      "LegacyTests\\": "legacy_tests/"
+    }
+  },
+  "extra": {
+    "laravel": {
+      "providers": [
+        "Livewire\\LivewireServiceProvider"
+      ],
+      "aliases": {
+        "Livewire": "Livewire\\Livewire"
+      }
+    }
+  },
+  "minimum-stability": "dev",
+  "prefer-stable": true
 }
