@@ -6,10 +6,10 @@ class MethodNotFoundException extends \Exception
 {
     use BypassViewHandler;
 
-    public function __construct($method, $component)
+    public function __construct($method)
     {
         parent::__construct(
-            "Unable to call component method. Public method [{$method}] not found on component: [{$component}]"
+            "Unable to call component method. Public method [{$method}] not found on component"
         );
     }
 }
