@@ -6,10 +6,12 @@ class MissingRulesException extends \Exception
 {
     use BypassViewHandler;
 
-    public function __construct($component)
+    public function __construct($instance)
     {
+        $class = $instance::class;
+
         parent::__construct(
-            "Missing [\$rules/rules()] property/method on Livewire component: [{$component}]."
+            "Missing [\$rules/rules()] property/method on: [{$class}]."
         );
     }
 }
